Added a 'draw_cards' method in the class 'Player' to draw cards from the top of the deck.
Added a 'choose_player' variable for testing purposes

Added a 'UnoMain' class. this will be in charge for how
the game starts n such, will change later on. All it does right now is
initailizes the game with a list of players,creates a deck, a discard pile,
and sets the first player and the direction of play.